pyproject.toml
[tool.poetry]
name = "nordigen-telegram-bot"
version = "0.1.0"
description = "Nordigen Telegram Bot"
authors = ["Viktor Sokolov"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^13.7"
requests = "^2.26.0"
SpeechRecognition = "^3.8.1"
pydub = "^0.25.1"
python-dotenv = "^0.19.0"
ffmpeg-python = "^0.2.0"
pip = "^21.2.4"

[tool.poetry.dev-dependencies]

[virtualenvs]
create = true
in-project = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
